move alertRuleRegistry to its own file
